refactor: single entry point to the components folder from where all components are exported
