+++ src/components/Main/Main.css
@@ -1,3 +1,3 @@
 .main {
-  background-color: aquamarine;
+  /* background-color: aquamarine; */
 }
