+++ src/components/index.js
@@ -1,0 +1,9 @@
+export { AsideLogout } from "./AsideLogout";
+export { AsideOption } from "./AsideOption";
+export { Avatar } from "./Avatar";
+export { Header } from "./Header";
+export { Main } from "./Main";
+export { NavBar } from "./NavBar";
+export { Search } from "./Search";
+export { SideBar } from "./SideBar";
+export { FieldSet } from "./FieldSet";
